+++ ziti/cmd/ziti/cmd/create_config.go
@@ -120,8 +120,7 @@ type RouterTemplateValues struct {
 	IsPrivate          bool
 	IsFabric           bool
 	IsWss              bool
-	TunnelerDisabled   bool
-	Tproxy             bool
+	TunnelerMode       string
 	IdentityCert       string
 	IdentityServerCert string
 	IdentityKey        string