+++ ziti/cmd/ziti/cmd/create_config_router.go
@@ -34,11 +34,11 @@ const (
 type CreateConfigRouterOptions struct {
 	CreateConfigOptions
 
-	RouterName       string
-	WssEnabled       bool
-	IsPrivate        bool
-	TunnelerDisabled bool
-	Tproxy           bool
+	RouterName   string
+	WssEnabled   bool
+	IsPrivate    bool
+	TunnelerMode string
+	LanInterface string
 }
 
 var routerOptions = CreateConfigRouterOptions{}